Inline test helper function
